pyproject.toml
[project]
name = "0-in-progress-fb-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.1",
    "gunicorn>=23.0.0",
    "openai>=1.97.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "torch>=2.7.1",
    "transformers>=4.54.0",
]
